Move split_tractogram to operations
