+++ scripts/scil_resample_tractogram.py
@@ -38,9 +38,10 @@ from scilpy.io.utils import (add_overwrite_arg, add_reference_arg,
                              add_verbose_arg,
                              assert_inputs_exist,
                              assert_outputs_exist)
-from scilpy.tracking.tools import (split_sft_randomly,
-                                   split_sft_randomly_per_cluster)
-from scilpy.tractograms.tractogram_operations import upsample_tractogram
+from scilpy.tractograms.tractogram_operations import (
+    split_sft_randomly,
+    split_sft_randomly_per_cluster,
+    upsample_tractogram)
 
 
 def _build_arg_parser():
